\name{add_to_database}
\alias{add_to_database}
\title{Adds a series of files to a database
Will automatically unzip files before calling them in}
\usage{
  add_to_database(db, files, table_name)
}
\arguments{
  \item{db}{a database connection object}

  \item{files}{a character vector of filenames to files to
  be imported}

  \item{table_name}{a name for the table to import to}
}
\description{
  Adds a series of files to a database Will automatically
  unzip files before calling them in
}

