% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical_codes}
\alias{clinical_codes}
\title{Clinical codes for 17 QOF conditions, smoking and HbA1c}
\format{
A dataframe with 1283 observations:
\describe{
  \item{medcode}{Unique internal code for the medical term selected by the healthcare professional}
  \item{readcode}{Unique Read Code for the medical term selected by the healthcare professional}
  \item{desc}{Description of the code}
  \item{list}{Name of the clinical code list this code belongs to}
}
}
\source{
\url{http://www.clinicalcodes.org}
}
\usage{
clinical_codes
}
\description{
This data comes from \url{http://www.clinicalcodes.org}
}
\keyword{datasets}
