% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cut_tv}
\alias{cut_tv}
\title{cut_tv - Cuts a survival dataset on a time varying variable}
\usage{
cut_tv(.data, entry, exit, cut_var, tv_name, cores = 1,
  on_existing = c("flip", "increment"))
}
\arguments{
\item{.data}{a dataframe}

\item{entry}{name of the column in .data that defines entry time to cohort. Column must be
numeric.}

\item{exit}{name of the column in .data that defines exit time from cohort.
Column must be numeric}

\item{cut_var}{name of the column in .data that defines the time of the time-varying covariate
event. Column must be numeric.}

\item{tv_name}{name for the constructed time-varying covariate}

\item{cores}{number of mc.cores to use.}

\item{on_existing}{see details for cutting behaviour}
}
\description{
Survival datasets often have time-varying covariates that need to be dealt with. For example
a drug exposure may occur after the entry into the cohort and you are interested in
how this might affect your outcome.
}
\details{
This function cuts up a dataset based on times supplied for the time-varying covariate. If there
is already a variable for the time-varying covariate, you can chose to flip the existing values
or increment them.  This means the function can be called multiple times to, e.g. deal with
drugs starting and stopping and also to deal with progression of treatment.

The function can also be parallelised for large datasets and chained with dply workflows.
arguments should not be quoted
}

