% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drop_all_temp_tables}
\alias{drop_all_temp_tables}
\title{drops all temporary tables from the database}
\usage{
drop_all_temp_tables(db)
}
\arguments{
\item{db}{a database connection}
}
\description{
This is useful for temporary storage/memory management
}

