% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ehr_def}
\alias{ehr_def}
\title{An example EHR_definition object for defining parameters for simulating EHR data}
\format{
object of type 'EHR_definition' including the following elements
\describe{
    \item{patient}{list of parameters defining the simulation of the patient table}
    \item{clinical}{list of parameters defining the simulation of the clinical table}
    \item{therapy}{list of parameters defining the simulation of the therapy table}
    \item{practice}{list of parameters defining the simulation of the patient table}
    \item{start_date}{start_date for patients - earliest possible birthday}
    \item{end_date}{Latest possible date in the ehr for data collection}
}
}
\usage{
ehr_def
}
\description{
An example EHR_definition object for defining parameters for simulating EHR data
}
\keyword{datasets}
