% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ehr_definition}
\alias{ehr_definition}
\title{An example EHR_definition object for defining parameters for simulating EHR data}
\format{object of type 'EHR_definition' including the following elements
\describe{
    \item{patient}{list of parameters defining the simulation of the patient table}
    \item{clinical}{list of parameters defining the simulation of the clinical table}
    \item{therapy}{list of parameters defining the simulation of the therapy table}
    \item{practice}{list of parameters defining the simulation of the patient table}
    \item{start_date}{start_date for patients - earliest possible birthday}
    \item{end_date}{Latest possible date in the ehr for data collection}
}}
\usage{
ehr_definition
}
\description{
An example EHR_definition object for defining parameters for simulating EHR data
}
\keyword{datasets}

