% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_string}
\alias{expand_string}
\title{Reads strings and expands sections wrapped in dotted parentheses}
\usage{
expand_string(s, level = 3)
}
\description{
This is a kind of inverse of bquote
 @export
 @param s a string
 @param level integer sets the parent frame level for evaluation
 @examples
 a <- runif(10)
 expand_string("The r code is .(a)")
}
