% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flat_files}
\alias{flat_files}
\title{Exports flat files from the database.  One file per practice}
\usage{
flat_files(
  db,
  table = "Consultation",
  practice_table = "Practice",
  out_dir,
  file_type = c("txt", "csv", "rda", "dta"),
  ...
)
}
\description{
@export
 @param db a database connection
 @param table character the table to be exported
 @param practice_table the table that the practice definitions can be found
 @param out_dir a directory to output to.  This will be created if it does not already exist
 @param file_type the type of file to be saved. This can be one of "txt", "csv", "rda", "dta".
 @param \dots arguments to be passed to export_fn
 
  Defaults to exporting consultation tables for use by \code{match_on_index()}.  the full path 
  to \code{out_dir} will be created if it does not already exist.
}
\seealso{
match_on_index
export_fn
}
