% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_system.R
\name{get_EHR_attribute}
\alias{get_EHR_attribute}
\title{Return the value of an attribute in the .ehr environment}
\usage{
get_EHR_attribute(x = NULL)
}
\arguments{
\item{x}{an attribute name}
}
\description{
Return the value of an attribute in the .ehr environment
}
\examples{
{
set_CPRD()
get_EHR_attribute()
get_EHR_attribute(patient_id)
}
}
