% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{import_definitions}
\alias{import_definitions}
\title{Imports definitions to be searched from a csv file into a MedicalDefinition object}
\usage{
import_definitions(input_file)
}
\arguments{
\item{input_file}{character path to the input file}
}
\description{
Imports definitions to be searched from a csv file into a MedicalDefinition object
}
\examples{
def2 <- import_definitions(system.file("extdata", "example_search.csv",
                                            package = "rpcdsearch"))
}
