% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{MedicalDefinition}
\alias{MedicalDefinition}
\title{Constructor function for MedicalDefinition class}
\usage{
MedicalDefinition(
  terms = NULL,
  codes = NULL,
  tests = NULL,
  drugs = NULL,
  drugcodes = NULL
)
}
\arguments{
\item{terms}{list of character vectors or NULL}

\item{codes}{list of character vectors or NULL}

\item{tests}{list of character vectors or NULL}

\item{drugs}{list of character vectors or NULL}

\item{drugcodes}{list of character vectors or NULL}
}
\description{
Constructor function for MedicalDefinition class
}
\details{
Elements marked with a "-" are excluded. Elements marked with a "r%" are resolved codes
}
\examples{
def <- MedicalDefinition(terms = list(c("angina", "unstable"), c("angina", "Crescendo "),
                                      c("angina", "Refractory")),
                         codes = list("G33..00", "G330.00", "\%r212H", "-G617"))
class(def)
}
