% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHR_definition.R
\name{print.EHR_definition}
\alias{print.EHR_definition}
\title{Tools for describing EMR_description objects.}
\usage{
\method{print}{EHR_definition}(x, element = NULL, level = 3, ...)
}
\arguments{
\item{x}{A \code{EHR_definition} object}

\item{element}{an element name}

\item{level}{nesting level for display purposes}

\item{\dots}{Additional arguments}
}
\description{
Tools for describing EMR_description objects.
}
