% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelists.R
\name{print.MedicalDefinition}
\alias{print.MedicalDefinition}
\title{Basic print method for medical definition classes}
\usage{
\method{print}{MedicalDefinition}(x, ...)
}
\arguments{
\item{x}{an object of class "medical_definition"}

\item{\dots}{Potential further arguments (required for method/generic reasons)}
}
\description{
Basic print method for medical definition classes
}
