% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{probabilistic_get_matches}
\alias{probabilistic_get_matches}
\title{Selected controls matching a list of variables from a case}
\usage{
probabilistic_get_matches(
  cases,
  control_pool,
  n_controls,
  match_vars,
  threshold = 0.5,
  cores = 1
)
}
\arguments{
\item{control_pool}{dataframe of potential controls for matching}

\item{n_controls}{The number of controls for each case. If replace == FALSE
this is a maximum value}

\item{matcher}{list of character strings defining the matching conditions}

\item{id}{named vector of length 1 for the variable name and value of the case identifier}

\item{replace}{logical should sampling of matched controls be with replacement (incident density
sampling) or not.}
}
\value{
A dataframe of matched controls or NULL if no controls could be found
}
\description{
Helper function for probabilistic_get_matches()
This function will perform probabilistic matching up to a supplied number 
of matched controls.
}
\examples{
\dontrun{
cases <- data.frame(patid = 1:3, name = c("Alice", "Bob", "Charlie"), age = c(25, 35, 45), gender = c("F", "M", "M"))
control_pool <- data.frame(patid = 4:10, name = c("Alicia", "Robert", "Charles", "Al", "Bobby", "Char", "Ali"), age = c(24, 36, 44, 26, 34, 46, 27), gender = c("F", "M", "M", "F"))
matched_controls <- probabilistic_get_matches(cases, control_pool, n_controls = 2, match_vars = c("name", "age", "gender"), threshold = 0.5, cores = 1)
print(matched_controls)
}
}
