% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{probabilistic_match_case}
\alias{probabilistic_match_case}
\title{Probabilistic match for a single case}
\usage{
probabilistic_match_case(
  case_row,
  control_pool,
  n_controls,
  match_vars,
  threshold = 0.5
)
}
\arguments{
\item{case_row}{A dataframe row representing a single case}

\item{control_pool}{A dataframe of potential controls}

\item{n_controls}{The number of controls to match to each case}

\item{match_vars}{A character vector of variable names to match on}

\item{threshold}{A numeric value for the minimum similarity score required}
}
\value{
A dataframe of matched controls or NULL if no controls could be found
}
\description{
Probabilistic match for a single case
}
