% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_by_year.R
\name{qof_years}
\alias{qof_years}
\title{Helper function providing startdate and enddate for a given year}
\usage{
qof_years(year)
}
\arguments{
\item{year}{integer}
}
\description{
Start and end dates matching QOF year start/ends
}
