% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_dates}
\alias{random_dates}
\title{Generates random dates between a start and end day
dates are in usual R as.Date() format}
\usage{
random_dates(n, start_day, end_day)
}
\arguments{
\item{n}{Number of dates to be returned}

\item{start_day}{string representation of a start day}

\item{end_day}{string representation of a start day}
}
\value{
a vector of dates
}
\description{
Enter start and end dates in ISO format, e.g. "2001-02-04"
}

