\name{read_zip}
\alias{read_zip}
\title{Reads a zipped data file to a dataframe}
\usage{
  read_zip(file, ...)
}
\arguments{
  \item{file}{character a file to read in}

  \item{...}{extra arguments to pass to read.delim}
}
\value{
  a dataframe
}
\description{
  Reads a zipped data file to a dataframe
}

