% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_system.R
\name{set_CPRD}
\alias{set_CPRD}
\title{Sets EHR metadata to CPRD format
When this is run, most functions in rEHR act as though the EHR database is CPRD}
\usage{
set_CPRD()
}
\description{
Sets EHR metadata to CPRD format
When this is run, most functions in rEHR act as though the EHR database is CPRD
}
\details{
Metadata on EHR type is stored in the .ehr environment. This allows the same functions to work 
across different data sources. The .ehr environment is not desgined to be accessible 
to the user, but accessor functions are provided. CPRD is the default EHR setting.
}
\seealso{
get_EHR_value
set_EHR_value
}
