% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_system.R
\name{set_EHR_attribute}
\alias{set_EHR_attribute}
\title{Sets the value of an attribute in the .ehr environment}
\usage{
set_EHR_attribute(x, value)
}
\arguments{
\item{x}{an ehr attribute name}

\item{value}{the value to set to the attribute}
}
\description{
Sets the value of an attribute in the .ehr environment
}
\examples{
set_CPRD()
set_EHR_attribute(practice_id, "pracid")
}
