% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_ehr_patients}
\alias{simulate_ehr_patients}
\title{Generate a dataframe of simulated patients with exit dates based on presented comorbidities.}
\usage{
simulate_ehr_patients(ehr_def)
}
\arguments{
\item{ehr_def}{an object of class ehr_def}
}
\value{
a dataframe of simulated patients
}
\description{
Generate a dataframe of simulated patients with exit dates based on presented comorbidities.
}
\details{
The definitions of the patient file are all in the ehr_def object
}

