% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_ehr_practices}
\alias{simulate_ehr_practices}
\title{generates a simulated dataframe of primary care practices in the same format as is used in CPRD}
\usage{
simulate_ehr_practices(ehr_def)
}
\arguments{
\item{ehr_def}{an object of class ehr_def}
}
\value{
a dataframe of simulated practices
}
\description{
generates a simulated dataframe of primary care practices in the same format as is used in CPRD
}
\details{
The definitions of the practice file are all in the ehr_def object
}

