% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_by_year.R
\name{standard_years}
\alias{standard_years}
\title{Helper function providing startdate and enddate for a given year}
\usage{
standard_years(year)
}
\arguments{
\item{year}{integer}
}
\description{
Standard years
}
