% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates}
\alias{convert_dates}
\title{converts date fields from ISO character string format to R Date format}
\usage{
convert_dates(dat, extras = NULL)
}
\arguments{
\item{dat}{a dataframe}

\item{extras}{= a character vector of extra columns to convert or NULL for no extras}
}
\description{
Date fields are determined by the date-fields element in the .ehr definition.  Extra
date fields can be added to the extras argument or by setting `.ehr$date_fields`.
}
\seealso{
get_EHR_value
set_EHR_value
}
