\name{database}
\alias{database}
\title{Wrapper for dbconnect
Connects to a SQLite database or creates one if it does not already exist}
\usage{
  database(dbname)
}
\arguments{
  \item{dbname}{a name for the new database}
}
\description{
  Wrapper for dbconnect Connects to a SQLite database or
  creates one if it does not already exist
}

