% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_tables.R
\name{drop_all_temp_tables}
\alias{drop_all_temp_tables}
\title{drops all temporary tables from the database}
\usage{
drop_all_temp_tables(db)
}
\arguments{
\item{db}{a database connection}
}
\description{
This is useful for temporary storage/memory management
}
