% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_string}
\alias{expand_string}
\title{Reads strings and expands sections wrapped in dotted parentheses}
\usage{
expand_string(s, level = 3)
}
\arguments{
\item{s}{a string}

\item{level}{integer sets the parent frame level for evaluation}
}
\description{
This is a kind of inverse of bquote
}
\examples{
a <- runif(10)
 expand_string("The r code is .(a)")
}

