% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prev_totals}
\alias{prev_totals}
\title{Calculates the prevalence totals for the output of a data frame of events/patients etc.}
\usage{
prev_totals(dat, included_totals = c("year", "practid"), time_var = "year",
  patient_var = "patid", original_data = TRUE)
}
\arguments{
\item{dat}{a dataframe}

\item{included_totals}{character vector describing which aggregates should be included e.g. c("year", "practid")}

\item{time_var}{name of the variable determining timepoints}

\item{patient_var}{name of the variable determining patient id for counting}

\item{original_data}{logical should the original data set be included in the output}
}
\value{
list of aggregates and/or the original data
}
\description{
e.g. Run on the output of a call to select_by_year
}
\details{
Outputs totals by aggregated by a time variable and other variables
Updated to use dplyr to do the aggregation - Now 40x faster!
}

