% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prev_totals}
\alias{prev_totals}
\title{Calculates the prevalence totals for the output of a data frame of events/patients etc.}
\usage{
prev_totals(dat, included_totals = c("patient", "practice"),
  original_data = TRUE, year_var = "year", patient_var = "patid",
  practice_var = "practid")
}
\arguments{
\item{dat}{a dataframe}

\item{included_totals}{character vector describing which aggregates should be included c("patient", "practice")}

\item{original_data}{logical should the original data set be included in the output}

\item{year_var}{name of the variable determining year}

\item{patient_var}{name of the variable determining patient id}

\item{practice_var}{name of the variable determining practice id}
}
\value{
list of aggregates and/or the original data
}
\description{
e.g. Run on the output of a call to select_by_year
}
\details{
Outputs totals by practice and/or by patient
}

