% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprd_import.R
\name{read_zip}
\alias{read_zip}
\title{Reads a zipped data file to a dataframe}
\usage{
read_zip(file, ...)
}
\arguments{
\item{file}{character a file to read in}

\item{\dots}{extra arguments to pass to read.delim}
}
\value{
a dataframe
}
\description{
This function will unzip a zipped text file and read it in to an R data frame
}
\details{
Default behaviour is to read in as a standard read.delim call.
extra arguments to read.delim can be passed to the function
}
